src/recommendation_service/requirements.txt
fastapi
uvicorn
python-jose[cryptography]
pyjwt
httpx
pinecone
pandas
numpy
scikit-learn
boto3 
pyarrow 
neo4j   
torch 
pymongo[srv]
redis
boto3
torch